+++ sm-shop-model/src/main/java/com/salesmanager/shop/model/catalog/product/attribute/ReadableProductOption3.java
@@ -13,10 +13,10 @@ import lombok.Setter;
 @Setter
 @JsonInclude(Include.NON_NULL)
 public class ReadableProductOption3 {
-	private int id=0;
+	private Long id=0L;
 	private String code="";
-	private int optionId=0;
-	private int descId=0;
+	private Long optionId=0L;
+	private Long descId=0L;
 	private int categoryId=0;
 	private String description="";
 	private List<ReadableProductOptionValue4> optionValue = new ArrayList<ReadableProductOptionValue4>();
